+++ lib/llm_composer/provider_stream_chunk/tensorx.ex
@@ -1,0 +1,7 @@
+defmodule LlmComposer.ProviderStreamChunk.TensorX do
+  @moduledoc false
+
+  use LlmComposer.ProviderStreamChunk.Struct,
+    parser: LlmComposer.ProviderStreamChunk.Parser.OpenAI,
+    provider: :tensorx
+end